+++ app/src/main/java/br/com/stone/emeraldcomponentsandroid/activities/LabelActivity.kt
@@ -1,0 +1,13 @@
+package br.com.stone.emeraldcomponentsandroid.activities
+
+import android.os.Bundle
+import android.support.v7.app.AppCompatActivity
+import br.com.stone.emeraldcomponentsandroid.R
+
+class LabelActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_label)
+    }
+}