+++ emeraldcomponents/src/main/java/br/com/stone/emeraldcomponents/basic/label/LabelSizeHandler.kt
@@ -1,0 +1,7 @@
+package br.com.stone.emeraldcomponents.basic.label
+
+import android.widget.TextView
+
+interface LabelSizeHandler {
+    fun setDimensions(labelText: TextView)
+}